+++ docs/NEWS.rst
@@ -46,6 +46,34 @@ Changes in HEAD (released ??/??/2020)
 	If you want to pass the variable's value instead as a simple IN parameter,
 	simply pass the local variable instead.
 
+*	In sisyphus jobs, when the value of the ``nextrun`` parameter was a
+	:class:`datetime.timedelta` object, it was interpreted relative to the
+	start of the run, not relative to now. This has been fixed.
+
+
+Changes in 5.64 (released 10/30/2020)
+-------------------------------------
+
+*	It is now possible to specify a custom port for ``ssh`` URLs.
+
+*	A second URL scheme ``ssh-nocheck`` has been added for ``ssh`` URLs.
+	Using ``ssh-nocheck`` disables the host key check when establishing the
+	``ssh`` connection.
+
+*	With these changes it is now possible to do the following::
+
+		>>> from ll import url
+		>>> u = url.URL("ssh-nocheck://www.example.org:2222/~/foo.txt")
+		>>> with url.Context():
+		... 	data = u.open("rb").read()
+
+
+Changes in 5.63.1 (released 10/26/2020)
+---------------------------------------
+
+*	Add a workaround for Python issue #41889 (https://bugs.python.org/issue41889)
+	to :class:`~ll.misc.Enum` and :class:`~ll.misc.IntEnum`.
+
 
 Changes in 5.63 (released 09/08/2020)
 -------------------------------------